refactor(api): Greatly simplify the handlers
Instead of having separate handlers for self-hosted and flagship
instances, roll them all into one. We can do this if we deserialize the
paths into a HashMap, and pick the parts we need, as needed. This way,
the `/:host` path can be made optional: if it isn't in the HashMap, we
default to the flagship instance, if any.
